add env-dev yaml file add tests add github actions for lint and test fix move impl in app module to separate app instantiation forbid extra keys in FPSConfig forbid overriding of pkg name add clear_names method on Config clear names before loading config add capability to use class plugins
