+++ .github/workflows/ci.yml
@@ -1,0 +1,27 @@
+name: CI
+
+on:
+  push:
+    branches:
+      - master
+  pull_request:
+    branches:
+      - master
+
+jobs:
+  test_quetz:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest]
+        python-version: ["3.7", "3.8", "3.9"]
+    steps:
+      - uses: actions/checkout@v2
+      - name: install micromamba
+        uses: mamba-org/provision-with-micromamba@main
+      - name: test FPS
+        shell: bash -l -eo pipefail {0}
+        run: |
+          micromamba install -f environment-dev.yml python=${{ matrix.python-version }}
+          pytest -vvv ./tests/
