+++ environment-dev.yml
@@ -1,0 +1,10 @@
+name: fps-dev
+channels:
+  - conda-forge
+dependencies:
+  - toml
+  - fastapi
+  - pluggy >=1.0,<2.0
+  - pytest
+  - pytest-mock
+  - pytest-lazy-fixture
